requirements.txt
streamlit
pillow
ultralytics
# torch 
# torchvision